Squashed 'libscsicmd/' changes from 19990bb..908fd2a
908fd2a Use dos2unix for consistent file content
23bd37c Adjust update-asc-num to subdir location
b298893 Move asc-num to structs subdir
96efd0c ninja build the test code
8ba1473 Fix warnings in build of test code
c3875ef Eliminate warnings
bd1b869 Switch to use ninja-build
80033ef Update ASC/ASCQ list
e00d585 Dump ATA info from sense in hex digits
93e4487 Implement ATA PASSTHROUGH 16 and use ATA_IDENTIFY with it
34227b6 Get subpages and correctly get all pages
c8f2a6e Add LOG SENSE command
b6f8855 Add translation from scsi device type to string
4ce4cbc Add test for SCSI INQUIRY command
32b8cc3 Add more attributes
a84d417 Fix a typo
d9ef47c Add attributes
63744c0 Implement a SMART attribute database
9cbb411 Parse ATA SMART READ THRESHOLD
53cdb59 Parse the ATA SMART READ DATA response
e199c16 Emit more info in sense_dump
cf2b272 Allow shorter inquiry responses
51109c0 Fix sign extension issue
2dc5943 Use an unsigned char buf as input

git-subtree-dir: libscsicmd
git-subtree-split: 908fd2a2294315029ec38fe54df7f9bf048ffff8

+++ src/parse_sense.c
@@ -170,7 +170,13 @@ static bool parse_sense_descriptor(unsigned char *sense, int sense_len, sense_in
                                         info->ata_status.error = sense[idx+3];
                                         if (info->ata_status.extend) {
                                                 info->ata_status.sector_count = (sense[idx+4] << 8) | sense[idx+5];
-                                                info->ata_status.lba = sense[idx+7] | (sense[idx+6]<<8) | (sense[idx+9]<<16) | (sense[idx+8]<<24) | ((uint64_t)sense[idx+11]<<32) | ((uint64_t)sense[idx+10]<<40);
+                                                info->ata_status.lba =
+													((uint64_t)sense[idx+7]) | 
+													((uint64_t)sense[idx+6]<<8) |
+													((uint64_t)sense[idx+9]<<16) |
+													((uint64_t)sense[idx+8]<<24) |
+													((uint64_t)sense[idx+11]<<32) |
+													((uint64_t)sense[idx+10]<<40);
                                         } else {
                                                 info->ata_status.sector_count = sense[idx+4];
                                                 info->ata_status.lba = sense[idx+7] | (sense[idx+9]<<8) | (sense[idx+11]<<16);
